requirements.txt
streamlit==1.31.0
pandas==2.2.0
Pillow==10.2.0
requests==2.31.0
pymongo==4.6.1
plotly==5.18.0
fpdf==1.7.2
